feat(公告): 添加富文本编辑器功能并优化公告详情页
refactor(导航): 将公告ID类型从Long改为Int
style(.gitignore): 添加RichEditorDemo到忽略列表
fix(公告详情): 增加加载状态处理并优化空ID检查

+++ composeApp/src/commonMain/kotlin/ovo/sypw/bsp/navigation/NavigationManager.kt
@@ -14,8 +14,8 @@ class NavigationManager {
     val currentScreen: State<String> = _currentScreen
 
     // 存储当前选中的公告ID
-    private val _selectedAnnouncementId = mutableStateOf<Long?>(null)
-    val selectedAnnouncementId: State<Long?> = _selectedAnnouncementId
+    private val _selectedAnnouncementId = mutableStateOf<Int?>(null)
+    val selectedAnnouncementId: State<Int?> = _selectedAnnouncementId
 
     /**
      * 导航到指定页面
@@ -29,7 +29,7 @@ class NavigationManager {
      * 导航到公告详情页面
      * @param announcementId 公告ID
      */
-    fun navigateToAnnouncementDetail(announcementId: Long) {
+    fun navigateToAnnouncementDetail(announcementId: Int) {
         _selectedAnnouncementId.value = announcementId
         _currentScreen.value = AppScreen.ANNOUNCEMENT_DETAIL.route
     }
